Перенес активацию карты в map.js
